requirements.txt
jupyter==1.1.1
numpy==2.2.4
matplotlib==3.10.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
pillow==11.1.0
soundfile==0.13.1
wandb==0.19.8
seaborn==0.11.2
